components/accelerator/nvidia/query/infiniband/parse_test.go
package infiniband

import (
	"os"
	"path/filepath"
	"testing"
)

func TestParseIBStat(t *testing.T) {
	input := `CA 'mlx5_0'
	CA type: MT4129
	Number of ports: 1
	Firmware version: 28.39.1002
	Hardware version: 0
	Node GUID: 0xa088c20300e3142a
	System image GUID: 0xa088c20300e3142a
	Port 1:
		State: Active
		Physical state: LinkUp
		Rate: 400
		Base lid: 0
		LMC: 0
		SM lid: 0
		Capability mask: 0x00010000
		Port GUID: 0xa288c2fffee3142a
		Link layer: Ethernet`

	parsed, err := ParseIBStat(input)
	if err != nil {
		t.Fatalf("Failed to parse ibstat output: %v", err)
	}
	t.Logf("parsed:\n%+v", parsed)
}

func TestParseIBStatFiles(t *testing.T) {
	files, err := filepath.Glob("testdata/ibstat.*")
	if err != nil {
		t.Fatalf("Failed to get ibstat files: %v", err)
	}
	for _, file := range files {
		t.Logf("file: %s", file)
		content, err := os.ReadFile(file)
		if err != nil {
			t.Fatalf("Failed to read file %s: %v", file, err)
		}
		parsed, err := ParseIBStat(string(content))
		if err != nil {
			t.Fatalf("Failed to parse ibstat file %s: %v", file, err)
		}
		t.Logf("parsed:\n%+v", parsed)
	}
}
